Add ValueRemover class that implements feature to remove values with certain Probability
